Complete feature reimplimentation
Everything now works as it did
